fix - Insufficient Modularization smells removed using extract class technique
